Monkey-patch around a crash when SG size > 128 WIs
cl_khr_sub_group_ballot seems to implicitly assume max 128
sized SGs due to the return value type uint4. However, the
max SG size is not stated anywhere in the specs and thus should
not affect the other sub group functionality to my understanding.

The 128 assumption is used for WI masking for the basic
subgroup test, causing a crash when exceeding it.

This just ups the limit to 1280 to push the limit up to
work around the issue. A proper fix would be to use dynamic bit
vector size here or define a max SG size in the specs (which
doesn't make sense).